# openapi.yaml
openapi: 3.0.3
info:
  title: OOH Audio Sync API
  description: Backend APIs for the DreamBricks Audio-Video Sync Totem system.
  version: "1.0.0"
servers:
  - url: http://localhost:8001
    description: Local Node.js Server
paths:
...
  /api/totems:
    get:
      summary: List Registered Totems
      description: Returns a list of all totems and their current configurations (online status, mobile count, and assigned video).
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "Totem_A"
                    is_online:
                      type: boolean
                      example: true
                    mobile_count:
                      type: integer
                      example: 3
                    video:
                      type: string
                      nullable: true
                      example: "ivete_video.mp4"
                    playlist:
                      type: array
                      description: Videos in order (loop), each with the audio phones play while it shows
                      items: { $ref: '#/components/schemas/PlaylistItem' }
                    schedule: { $ref: '#/components/schemas/Schedule' }
                    showing:
                      type: string
                      nullable: true
                      description: Campaign whose content is on air now (itself, its fallback, or null = black screen with logo)
    post:
      summary: Create Totem
      description: ID must be 1–40 letters, numbers, `-` or `_`. Video/audio are optional but must exist in `assets/`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, example: shopping_paulista_1 }
                video: { type: string, example: promo.mp4 }
                audio: { type: string, example: promo.mp3 }
                playlist:
                  type: array
                  description: Instead of video/audio, several videos in order (max 20)
                  items: { $ref: '#/components/schemas/PlaylistItem' }
                schedule: { $ref: '#/components/schemas/Schedule' }
      responses:
        '201':
          description: Created
        '400':
          description: Invalid ID or media not found
        '409':
          description: ID already exists
  /api/totem/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        example: totem1
    patch:
      summary: Update or Rename Totem
      description: |
        Updates video/audio and/or renames the totem (`id`). Settings and mobile links move to the new ID.
        If the totem screen is online it receives `change_screen` and reloads with `?screen=<new id>`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, example: totem1_lobby }
                video: { type: string }
                audio: { type: string }
                playlist:
                  type: array
                  items: { $ref: '#/components/schemas/PlaylistItem' }
                schedule:
                  allOf: [{ $ref: '#/components/schemas/Schedule' }]
                  description: "null = always on air"
      responses:
        '200':
          description: Updated
        '400':
          description: Invalid ID or media not found
        '404':
          description: Totem not found
        '409':
          description: New ID already exists
    delete:
      summary: Delete Totem
      description: Removes the saved settings (video, audio, mobile links). `still_online` tells if its screen is still connected.
      responses:
        '200':
          description: Deleted
        '404':
          description: Totem not found
...
  /api/totem/{id}/promo:
    put:
      summary: Update Mobile Page Text and Links
      description: |
        Sets what the mobile (audio) page shows for this totem: a message, the app download
        button (store link per device) and up to 6 link buttons. URLs must be http(s).
        Totems never edited use the default (99food) content.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: totem1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string, example: "Enquanto escuta, saiba mais sobre a marca" }
                app:
                  type: object
                  properties:
                    label: { type: string, example: Baixar App }
                    ios: { type: string, example: "https://apps.apple.com/..." }
                    android: { type: string, example: "https://play.google.com/..." }
                    fallback: { type: string, example: "https://example.com" }
                links:
                  type: array
                  items:
                    type: object
                    properties:
                      label: { type: string, example: Site }
                      url: { type: string, example: "https://example.com" }
                      icon: { type: string, example: globe }
      responses:
        '200':
          description: Saved
        '400':
          description: Invalid URL, missing label or too many links
...
  /api/totem/{id}/config:
    post:
      summary: Set the Videos of a Totem
      description: |
        Sets what a campaign plays: `{ playlist: [{ video, audio }] }` (videos in order, loop) or the
        single `{ video, audio }`. Every item needs a video and an audio here. Open screens switch right away.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          example: totem1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                video: { type: string, example: promo.mp4 }
                audio: { type: string, example: promo.mp3 }
                playlist:
                  type: array
                  items: { $ref: "#/components/schemas/PlaylistItem" }
      responses:
        "200":
          description: "`{ success, id, video, audio, playlist }`"
        "400":
          description: Missing video/audio or file not in the library
...
components:
  schemas:
    PlaylistItem:
      type: object
      properties:
        video: { type: string, example: verao_video.mp4 }
        audio: { type: string, example: verao_audio.mp3, description: "Empty = silence on the phone during this video" }
    Schedule:
      type: object
      nullable: true
      description: Period on air (UTC ISO dates, both optional). Outside it the campaign shows `fallback`'s content, or a black screen with the logo.
      properties:
        start: { type: string, format: date-time, nullable: true }
        end: { type: string, format: date-time, nullable: true }
        fallback: { type: string, nullable: true, description: Another campaign id }